# runtime/openapi.yaml
openapi: 3.1.0
info:
  title: PRAXIOS Runtime API
  version: 0.2.0
  description: Governed execution API for PRAXIOS, Meta-Harness and Decision Room.
servers:
  - url: https://runtime.example.com
paths:
...
  /api/providers:
    get:
      summary: List configured provider adapters
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Provider registry }
        '401': { description: Unauthorized }
  /api/sessions:
    post:
      summary: Create a governed session
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [goal]
              properties:
                sessionId: { type: string }
                goal: { type: string }
                actor: { type: string }
      responses:
        '201': { description: Session snapshot }
...
  /api/orchestrate:
    post:
      summary: Run a planner-worker-verifier governed session
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrchestrationRequest'
      responses:
        '200': { description: Governed orchestration result }
        '400': { description: Contract failure }
        '429': { description: Budget or rate limit exceeded }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  schemas:
...
    ModelSlot:
      type: object
      required: [provider, model]
      properties:
        provider: { type: string }
        model: { type: string }
        actorId: { type: string }
    OrchestrationRequest:
      type: object
      required: [goal, planner, workers, verifier]
      properties:
        goal: { type: string }
        actor: { type: string }
        planner: { $ref: '#/components/schemas/ModelSlot' }
        verifier: { $ref: '#/components/schemas/ModelSlot' }
        workers:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/ModelSlot'
        budget:
          type: object
          properties:
            maxProviderCalls: { type: integer }
            maxTasks: { type: integer }
            maxInputChars: { type: integer }
            maxOutputChars: { type: integer }
            maxWallClockMs: { type: integer }